requirements.txt
fastapi==0.110.1
PyJWT[crypto]==2.8.0
httpx>=0.28.1,<1.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
cryptography==42.0.8
osfclient @ git+https://github.com/RCOSDP/rdmclient.git
